Created a seervice for adding missing persons to the database
