+++ server/internal/service/validate_person.go
@@ -3,11 +3,10 @@ package service
 import (
 	"fmt"
 	"missing-persons-backend/internal/models"
-	"missing-persons-backend/repository"
+	"missing-persons-backend/internal/repository"
 	"regexp"
 	"time"
 
-	"github.com/google/uuid"
 )
 
 type PersonService struct {
@@ -16,29 +15,7 @@ type PersonService struct {
 }
 
 
-func (s*PersonService) CreatePerson ( person models.Person) error {
 
-
-	person.ID =uuid.New().String()
-	err := ValidatePerson(person)
-
-	if err != nil {
-
-
-		return fmt.Errorf("validation failed : %w", err)
-
-	}
-
-
-	err = s.Repo.CreatePerson(person)
-
-	if err != nil {
-
-		return fmt.Errorf("failed to create person : %w",err)
-	}
-
-	return nil
-}
 
 
 func ValidatePerson(person models.Person) error {
